% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_pal.R
\name{name of palette to retrieve}
\alias{name of palette to retrieve}
\alias{amino_acid_pal}
\title{Provide common color palettes for amino acids}
\usage{
amino_acid_pal(name)
}
\description{
Provide common color palettes for amino acids
}
\examples{
amino_acid_pal('shapely')
#' @export
}
\keyword{amino-acid}
\keyword{color}
\keyword{protein}
